Rename packet module to plural
