+++ mqtt-format/src/v5/mod.rs
@@ -2,7 +2,7 @@ pub mod bytes;
 pub mod fixed_header;
 pub mod integers;
 pub mod level;
-pub mod packet;
+pub mod packets;
 pub mod properties;
 pub mod reason_code;
 pub mod strings;
